Win32 port changes

+++ source/server/guarddog_impl.h
@@ -43,7 +43,7 @@ public:
   /**
    * Exposed for testing purposes only (but harmless to call):
    */
-  int loopIntervalForTest() const { return loop_interval_.count(); }
+  long long loopIntervalForTest() const { return loop_interval_.count(); }
   void forceCheckForTest() {
     exit_event_.notifyAll();
     Thread::LockGuard guard(exit_lock_);
